Add new Form and Watches components, refactor page.tsx
Added new Form and Watches components to help manage state better in the application. These components help create a Form and manage the state of Watches in a more structured way. Also refactored 'page.tsx' to use these newly added components.

In the previous iteration, there were no separate components to handle the state of the Form and Watches which made the 'page.tsx' cluttered and difficult to maintain.

By splitting into separate components, it has become easier to manage the state. It also makes our code cleaner and adheres better to the single responsibility principle. The Form component is used to dispatch 'add' and 'remove' actions and the Watches component is used to perform the respective tasks.

CSS was also updated to style these new components and some redundant CSS was removed. Also, replaced Inter font with Roboto to provide a better look and feel to the application.